+++ axum/src/extract/matched_path.rs
@@ -296,6 +296,27 @@ mod tests {
     }
 
     #[crate::test]
+    async fn can_extract_nested_matched_path_with_prefix_and_suffix_in_middleware_on_nested_router()
+    {
+        async fn extract_matched_path<B>(matched_path: MatchedPath, req: Request<B>) -> Request<B> {
+            assert_eq!(matched_path.as_str(), "/f{o}o/b{a}r");
+            req
+        }
+
+        let app = Router::new().nest(
+            "/f{o}o",
+            Router::new()
+                .route("/b{a}r", get(|| async move {}))
+                .layer(map_request(extract_matched_path)),
+        );
+
+        let client = TestClient::new(app);
+
+        let res = client.get("/foo/bar").await;
+        assert_eq!(res.status(), StatusCode::OK);
+    }
+
+    #[crate::test]
     async fn can_extract_nested_matched_path_in_middleware_on_nested_router_via_extension() {
         async fn extract_matched_path<B>(req: Request<B>) -> Request<B> {
             let matched_path = req.extensions().get::<MatchedPath>().unwrap();