+++ axum/CHANGELOG.md
@@ -42,6 +42,10 @@ Yanked from crates.io due to unforeseen breaking change, see [#3190] for details
 
 - **fixed:** Removed the warning about breaking changes from README
 
+- **changed:** Updated `matchit` allowing for routes with captures and static prefixes and suffixes ([#3143])
+
+[#3143]: https://github.com/tokio-rs/axum/pull/3143
+
 # 0.8.0
 
 ## since rc.1