+++ axum/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- **changed:** `serve` has an additional generic argument and can now work with any response body
+  type, not just `axum::body::Body` ([#3205])
+
+[#3205]: https://github.com/tokio-rs/axum/pull/3205
+
 # 0.8.4
 
 - **added:** `Router::reset_fallback` ([#3320])