+++ axum/CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+- **breaking:** Router fallbacks are now properly merged for nested routers ([#3158])
 - **added:** Implement `OptionalFromRequest` for `Multipart` ([#3220])
 - **changed:** `serve` has an additional generic argument and can now work with any response body
   type, not just `axum::body::Body` ([#3205])
 
+[#3158]: https://github.com/tokio-rs/axum/pull/3158
 [#3205]: https://github.com/tokio-rs/axum/pull/3205
+[#3220]: https://github.com/tokio-rs/axum/pull/3220
 
 # 0.8.4