+++ axum/CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- **breaking:** Router fallbacks are now properly merged for nested routers ([#3158])
+- **added:** Implement `OptionalFromRequest` for `Multipart` ([#3220])
+- **changed:** `serve` has an additional generic argument and can now work with any response body
+  type, not just `axum::body::Body` ([#3205])
+
+[#3158]: https://github.com/tokio-rs/axum/pull/3158
+[#3205]: https://github.com/tokio-rs/axum/pull/3205
+[#3220]: https://github.com/tokio-rs/axum/pull/3220
+
+# 0.8.4
+
+- **added:** `Router::reset_fallback` ([#3320])
+- **added:** `WebSocketUpgrade::selected_protocol` ([#3248])
+- **fixed:** Panic location for overlapping method routes ([#3319])
+- **fixed:** Don't leak a tokio task when using `serve` without graceful shutdown ([#3129])
+
+[#3319]: https://github.com/tokio-rs/axum/pull/3319
+[#3320]: https://github.com/tokio-rs/axum/pull/3320
+[#3248]: https://github.com/tokio-rs/axum/pull/3248
+[#3129]: https://github.com/tokio-rs/axum/pull/3129
+
 # 0.8.3
 
 - **added:** Implement `From<Bytes>` for `Message` ([#3273])